view: add a `node id -> idx` map for the current nodes
This is likely faster than `find` on `List<T>` (removed in this commit
as it was only used for the current node list) which does binary search.

The real reason we want this map is to modify the current selection;
many operations (like `neighbors` and `subgraph`) will produce a NodeId
list (or really a set) which we'll then need to translate into list
indexes to update our selection.

Note that recording our selection as a node id set is not a solution; we
have operations that require getting node ids out of our selection:
  - communicating in the UI which nodes are selected
  - finding the id of the closest selected node to the cursor

In other words, we have a need for both fast `idx -> node id` lookup and
`node id -> idx` lookup. The current list vector fulfills the first need
and the map added in this commit does the latter.