command-table: allow `post_parse_hook` to have an independent return type
we want this for cases where the command wants to produce data that's
out of band with the subcommands

`SelectionCommand` is an example of this; `op` is a property all
subcommands have

Note: it'd be nicer if we could get clap to parse `op` for us (it does
have support for "global" args) but a cursory exploration didn't reveal
any easy way to get it to do the kind of oddball "first character"
parsing that we're doing for the op. Additionally, we'd have to
implement `Subcommand` on the `SelectionCommand` struct ourselves which
is... annoying.