action: add `mk-umbrella` (aka `mk-grouped`)
see comments/docs within for details; in short:

converts a graph like this:
```graphviz
digraph {
    a -> b;
    b;
    c -> a;
    d;

    target1 -> a;
    target1 -> b;
    target1 -> c;
    target1 -> d;

    target2 -> a;
    target2 -> b;
    target2 -> c;
    target2 -> d;

    target3 -> a;
    target3 -> b;
    target3 -> c;
    target3 -> d;

    target4 -> a;
    target4 -> b;
    target4 -> c;
    target4 -> d;
}
```

to this:
```graphviz
digraph {
	a -> b;
	b;
	c -> a;
	d;

	target1;
	target2;
	target3;
	target4;

	srcs [
			peripheries="2"
			style="filled"
			shape="invhouse"
			fillcolor="#F0387860"
			label="srcs (4 nodes)

(umbrella: select({ c } | { d } | { a } | { b }))
(net edges: -8)"
	];

	srcs -> a
	srcs -> b
	srcs -> c
	srcs -> d

	target1 -> srcs
	target2 -> srcs
	target3 -> srcs
	target4 -> srcs
}
```

when asked to put an umbrella over `{ a, b, c, d }`

For build graphs that declare lots of targets (i.e. with a static
pattern in GNU Make) that all depend on a common set of things, this
(umbrellas) can be extremely useful in cleaning up the graph and making
it more legible. Ideally the common set of things would already be
factored out into their own group/placeholder target but this isn't
always the case.